backend/pyproject.toml
[tool.poetry]
name = "socialhack-backend"
version = "0.1.0"
description = "SocialHack backend dependencies"
authors = ["BilliardBytes <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-cors-headers = "^4.2.0"
django-environ = "^0.10.0"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-stubs-ext = "^0.7.0"
django-split-settings = "^1.2.0"
dj-rest-auth = "^4.0.1"
drf-yasg = "^1.21.6"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
weasyprint = "^52.5"
django-debug-toolbar = "^4.1.0"


[build-system]
# Should be the same as `$POETRY_VERSION`
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
